Fixed serial timeout overflow

+++ Arduino/LEDstream_FastLED/LEDstream_FastLED.ino
@@ -239,7 +239,7 @@ void timeouts(){
 		lastAckTime = t; // Reset counter
 
 		// If no data received for an extended time, turn off all LEDs.
-		if((t - lastByteTime) >= SerialTimeout * 1000) {
+		if((t - lastByteTime) >= (uint32_t) SerialTimeout * 1000) {
 			memset(leds, 0, Num_Leds * sizeof(struct CRGB)); //filling Led array by zeroes
 			FastLED.show();
 			mode = Header;
